# perf(langgraph): Speed up method `PregelLoop._emit` by 243%