@casparb 2025-09-19
Hi @misrasaurabh1, declining this. You've introduced a behavior change by batching a single timestamp instead of per-item timestamps. It also includes promo, non-reproducible perf claims, and was opened by a marketing bot. Please open an issue with a reproducible benchmark and evidence that _emit is a real bottleneck before proposing risky micro optimizations!